♻️ refactor : Station RestAssured 코드 클래스로 분리
